fix-readBody

+++ includes/RequestReader.hpp
@@ -52,7 +52,7 @@ class RequestReader
 		void								readMultipartInfo(const std::string& boundary, std::vector<char> &tempLine);
 
 		// READLINE AND UTILS
-		void	 							readAllContentLength(int fd, int contentLength);
+		void	 							readAllContentLength(int fd, long int contentLength);
 		void								readUntilSize(int fd, long int size);
 		void								readUntilCRLF(int fd, std::string &segment);
 		std::string							intToString(int value);
